added disqus comments. updated post_generator

+++ post_generator.py
@@ -7,9 +7,12 @@ TEXT = '''---
 layout: {{layout}}
 title: {{title}}
 author: {{author}}
+comments: {{use_comments}}
 ---
 
+<!--========================== Blog =========================-->
 Enter blog text here
+<!--=========================== EOF =========================-->
 '''
 
 
@@ -23,15 +26,26 @@ def generate_post(args):
         return
 
     temp = Template(TEXT)
-    author = args.author
-    layout = args.layout
+
+    print 'Rendering template'
+    for ai in dir(args):
+        if not ai.startswith('_'):
+            print '    {:<15s}= {}'.format(ai, getattr(args, ai))
+
     txt = temp.render(title=title,
-                      author=author,
-                      layout=layout)
+                      author=args.author,
+                      layout=args.layout,
+                      use_comments=not args.nocomments)
 
-    with open(p, 'w') as fp:
-        fp.write(txt)
+    print 'Generated Blog boilerplate\n'
+    print txt
 
+    if not args.dryrun:
+        print 'Writing blog boilerplate to {}'.format(p)
+        with open(p, 'w') as fp:
+            fp.write(txt)
+    else:
+        print 'Dry Run. Not writing boilerplate'
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Make a pychron application')
@@ -47,9 +61,16 @@ if __name__ == '__main__':
                         type=str,
                         default='post',
                         help='set layout for this post')
+
+    parser.add_argument('-n', '--nocomments',
+                        action="store_true",
+                        help='Do not use Disqus comments for this post')
+    parser.add_argument('-d','--dryrun',
+                        action="store_true",
+                        help='Do a dry run. Do not write file.')
     # if args:
     # title = args[0]
     # else:
-    #     title = 'blank'
+    # title = 'blank'
     args = parser.parse_args()
     generate_post(args)
